+++ @angular-generic-table/core/pipes/gt-is-observable.pipe.ts
@@ -1,0 +1,11 @@
+import { Pipe, PipeTransform } from '@angular/core';
+import {Observable} from 'rxjs/Observable';
+
+@Pipe({
+  name: 'gtIsObservable'
+})
+export class GtIsObservablePipe implements PipeTransform {
+  transform(input: any): boolean {
+    return input instanceof Observable;
+  }
+}